Address utility ownership and type annotation feedback
